feat: update code and add TODO
