+++ src/main/java/com/blogify/blogapi/repository/CommentReactRepository.java
@@ -1,6 +1,6 @@
 package com.blogify.blogapi.repository;
 
-import com.blogify.blogapi.repository.model.CommentReact;
+import com.blogify.blogapi.model.CommentReact;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
